start exit monitor after fills
